+++ config/routes.rb
@@ -1,5 +1,8 @@
 KyufyCore::Engine.routes.draw do
   resources :assessments, only: [ :create ] do
-    post :batch, on: :collection
+    collection do
+      post :batch
+      post :household
+    end
   end
 end